rpc added

+++ karmanode_setup.sh
@@ -20,6 +20,8 @@ function gather_user_info () {
 	read -p "  Karmanode outputs tx_hash: " KARMANODE_TX_HASH
 	read -p "  Karmanode outputs idx: " KARMANODE_IDX
 	read -p "  Karmanode user (Default ohmcoin): " KARMANODE_USER
+	read -p "  Karmanode rpc user: " KARMANODE_RPC_USER
+	read -p "  Karmanode rpc password: " KARMANODE_RPC_PASS
 
 	if [ -z "$KARMANODE_GENKEY" ]; then
 		echo "=> Information provided on Karmanode genkey is not correct."
@@ -31,6 +33,14 @@ function gather_user_info () {
 	fi
 	if [ -z "$KARMANODE_IDX" ]; then
 		echo "=> Information provided on Karmanode outputs idx is not correct."
+		ERRORS=true
+	fi
+	if [ -z "$KARMANODE_RPC_USER" ]; then
+		echo "=> Information provided on Karmanode rpc user is not correct."
+		ERRORS=true
+	fi
+	if [ -z "$KARMANODE_RPC_PASS" ]; then
+		echo "=> Information provided on Karmanode rpc password is not correct."
 		ERRORS=true
 	fi
 	if [ -z "$KARMANODE_USER" ]; then
@@ -51,6 +61,8 @@ function check_info () {
 	echo "Karmanode outputs tx_hash: $KARMANODE_TX_HASH"
 	echo "Karmanode outputs idx: $KARMANODE_IDX"
 	echo "Karmanode user: $KARMANODE_USER"
+	echo "Karmanode rpc user: $KARMANODE_RPC_USER"
+	echo "Karmanode rpc password: $KARMANODE_RPC_PASS"
 
 	read -p "Procced? [y/n]: " PROCCED
 	if [ "$PROCCED" != "y" ]; then
@@ -81,13 +93,18 @@ function user_creation () {
 
 function modify_sshd_config () {
 
-	cp /etc/ssh/sshd_config /etc/ssh/sshd_config_$DATE
-	sed -re 's/^(\#)(PasswordAuthentication)([[:space:]]+)(.*)/\2\3\4/' -i /etc/ssh/sshd_config
-	sed -re 's/^(\#?)(PasswordAuthentication)([[:space:]]+)yes/\2\3no/' -i /etc/ssh/sshd_config
-	sed -re 's/^(\#)(UsePAM)([[:space:]]+)(.*)/\2\3\4/' -i /etc/ssh/sshd_config
-	sed -re 's/^(\#?)(UsePAM)([[:space:]]+)yes/\2\3no/' -i /etc/ssh/sshd_config
-	sed -re 's/^(\#)(PermitRootLogin)([[:space:]]+)(.*)/\2\3\4/' -i /etc/ssh/sshd_config
-	sed -re 's/^(\#?)(PermitRootLogin)([[:space:]]+)(.*)/\2\3no/' -i /etc/ssh/sshd_config
+	if [ -f /etc/ssh/sshd_config ]; then
+		cp /etc/ssh/sshd_config /etc/ssh/sshd_config_$DATE
+		sed -re 's/^(\#)(PasswordAuthentication)([[:space:]]+)(.*)/\2\3\4/' -i /etc/ssh/sshd_config
+		sed -re 's/^(\#?)(PasswordAuthentication)([[:space:]]+)yes/\2\3no/' -i /etc/ssh/sshd_config
+		sed -re 's/^(\#)(UsePAM)([[:space:]]+)(.*)/\2\3\4/' -i /etc/ssh/sshd_config
+		sed -re 's/^(\#?)(UsePAM)([[:space:]]+)yes/\2\3no/' -i /etc/ssh/sshd_config
+		sed -re 's/^(\#)(PermitRootLogin)([[:space:]]+)(.*)/\2\3\4/' -i /etc/ssh/sshd_config
+		sed -re 's/^(\#?)(PermitRootLogin)([[:space:]]+)(.*)/\2\3no/' -i /etc/ssh/sshd_config
+	else
+		echo "/etc/ssh/sshd_config does not exist, please review your ssh daemon configuration and try again. Aborting."
+		exit
+	fi
 
 }
 
@@ -121,15 +138,26 @@ echo "--- Creating swap partition..."
 create_swap_space
 
 echo "Installing ohmc daemon..."
-su ohmcoin
-cd
-git clone https://github.com/theohmproject/ohmcoin.git 
-cd ohmcoin 
-chmod +x share/genbuild.sh 
-chmod +x autogen.sh 
-chmod 755 src/leveldb/build_detect_platform
+
+su -c "cd; git clone https://github.com/theohmproject/ohmcoin.git" $KARMANODE_USER
+su -c "cd ~/ohmcoin; chmod +x share/genbuild.sh" $KARMANODE_USER
+su -c "cd ~/ohmcoin; chmod +x autogen.sh" $KARMANODE_USER
+su -c "cd ~/ohmcoin; chmod 755 src/leveldb/build_detect_platform" $KARMANODE_USER
+
 #sudo ./autogen.sh 
 #sudo ./configure
 #sudo make
 #sudo make install
+
 echo "Installing blockchain snapshot..."
+
+echo "Configuring ohmc daemon..."
+
+#mkdir ~/.ohmc/ && touch ~/.ohmc/ohmc.conf
+#echo "rpcuser=any_user" >> ~/.ohmc/ohmc.conf
+#echo "rpcpassword=any_password" >> ~/.ohmc/ohmc.conf
+
+#cd ~/.ohmc
+#ohmcd -daemon
+
+echo "Waiting to daemon to get synced..."
